+++ src/apps/seelenweg/events.ts
@@ -1,7 +1,6 @@
 import { toPhysicalPixels } from '../utils';
 import { SeelenWegHideMode, SeelenWegSide } from '../utils/schemas/Seelenweg';
 import { debounce, TimeoutIdRef } from '../utils/Timing';
-import { PhysicalSize } from '@tauri-apps/api/dpi';
 import { emitTo } from '@tauri-apps/api/event';
 import { getCurrent } from '@tauri-apps/api/webviewWindow';
 
@@ -14,26 +13,36 @@ const root_container = document.getElementById('root')!;
 export const ExtraCallbacksOnLeave = new CallbacksManager();
 export const ExtraCallbacksOnActivate = new CallbacksManager();
 
-export const setWindowSize = () => {
-  const webview = getCurrent();
-  const screenWidth = Math.floor(window.screen.width * window.devicePixelRatio);
-  const screenHeight = Math.floor(window.screen.height * window.devicePixelRatio);
-  webview.setSize(new PhysicalSize(screenWidth, screenHeight));
-};
+export const updateHitbox = debounce(async () => {
+  const {
+    isOverlaped,
+    settings: { position, hideMode },
+  } = store.getState();
+  const view = getCurrent();
+  const hitboxTarget = view.label.replace('/', '-hitbox/');
 
-export const updateHitbox = debounce(() => {
-  const { isOverlaped, settings: { position, hideMode } } = store.getState();
+  const viewPosition = await view.innerPosition();
+  const viewSize = await view.innerSize();
 
-  const isAutoHideOn = (hideMode !== SeelenWegHideMode.Never && isOverlaped) || hideMode === SeelenWegHideMode.Always;
+  const isAutoHideOn =
+    (hideMode !== SeelenWegHideMode.Never && isOverlaped) || hideMode === SeelenWegHideMode.Always;
   const isHorizontal = position === SeelenWegSide.TOP || position === SeelenWegSide.BOTTOM;
 
-  const hiddenOffsetTop = position === SeelenWegSide.TOP ? 0 : toPhysicalPixels(window.screen.height) - 1;
-  const hiddenOffsetLeft = position === SeelenWegSide.LEFT ? 0 : toPhysicalPixels(window.screen.width) - 1;
-  emitTo('seelenweg-hitbox', 'move', {
-    x: isAutoHideOn && !isHorizontal ? hiddenOffsetLeft : toPhysicalPixels(root_container.offsetLeft),
-    y: isAutoHideOn && isHorizontal ? hiddenOffsetTop : toPhysicalPixels(root_container.offsetTop),
+  const hiddenOffsetTop =
+    position === SeelenWegSide.TOP ? viewPosition.y : viewPosition.y + viewSize.height - 1;
+  const hiddenOffsetLeft =
+    position === SeelenWegSide.LEFT ? viewPosition.x : viewPosition.x + viewSize.width - 1;
+  emitTo(hitboxTarget, 'move', {
+    x:
+      isAutoHideOn && !isHorizontal
+        ? hiddenOffsetLeft
+        : viewPosition.x + toPhysicalPixels(root_container.offsetLeft),
+    y:
+      isAutoHideOn && isHorizontal
+        ? hiddenOffsetTop
+        : viewPosition.y + toPhysicalPixels(root_container.offsetTop),
   });
-  emitTo('seelenweg-hitbox', 'resize', {
+  emitTo(hitboxTarget, 'resize', {
     width: isAutoHideOn && !isHorizontal ? 1 : toPhysicalPixels(root_container.offsetWidth),
     height: isAutoHideOn && isHorizontal ? 1 : toPhysicalPixels(root_container.offsetHeight),
   });
@@ -43,11 +52,15 @@ export function registerDocumentEvents() {
   const timeoutId: TimeoutIdRef = { current: null };
   const webview = getCurrent();
 
-  const onMouseLeave = debounce(() => {
-    webview.setIgnoreCursorEvents(true);
-    ExtraCallbacksOnLeave.execute();
-    updateHitbox();
-  }, 200, timeoutId);
+  const onMouseLeave = debounce(
+    () => {
+      webview.setIgnoreCursorEvents(true);
+      ExtraCallbacksOnLeave.execute();
+      updateHitbox();
+    },
+    200,
+    timeoutId,
+  );
 
   const onMouseEnter = () => {
     if (timeoutId.current) {