+++ src/apps/seelenweg/index.tsx
@@ -1,9 +1,8 @@
 import { getRootContainer } from '../utils';
 import { wrapConsole } from '../utils/ConsoleWrapper';
 import { ErrorBoundary } from './components/Error';
-import { registerDocumentEvents, setWindowSize, updateHitbox } from './events';
+import { registerDocumentEvents, updateHitbox } from './events';
 import { SeelenWeg } from './modules/bar';
-import { invoke } from '@tauri-apps/api/core';
 import { emitTo } from '@tauri-apps/api/event';
 import { getCurrent } from '@tauri-apps/api/webviewWindow';
 import { ConfigProvider, theme } from 'antd';
@@ -19,24 +18,27 @@ import './styles/variables.css';
 import './styles/reset.css';
 import './styles/global.css';
 
+async function onMount() {
+  let view = getCurrent();
+  updateHitbox();
+  await emitTo(view.label.replace('/', '-hitbox/'), 'init');
+  await view.show();
+  await view.emit('complete-setup');
+}
+
 async function Main() {
   wrapConsole();
   const container = getRootContainer();
 
-  setWindowSize();
   registerDocumentEvents();
 
   await loadConstants();
   await loadStore();
   await registerStoreEvents();
-  await invoke('enum_opened_apps');
 
   const WrappedRoot = () => {
     useEffect(() => {
-      emitTo('seelenweg-hitbox', 'init').then(() => {
-        updateHitbox();
-        getCurrent().show();
-      });
+      onMount();
     }, []);
 
     return (