+++ src/apps/toolbar/app.tsx
@@ -1,5 +1,4 @@
 import { ErrorBoundary } from '../seelenweg/components/Error';
-import { toPhysicalPixels } from '../utils';
 import { ErrorFallback } from './components/Error';
 import { emitTo } from '@tauri-apps/api/event';
 import { getCurrent } from '@tauri-apps/api/webviewWindow';
@@ -10,19 +9,17 @@ import { ToolBar } from './modules/main/infra';
 
 import { Selectors } from './modules/shared/store/app';
 
-async function onMount(height: number) {
+async function onMount() {
   let view = getCurrent();
   await emitTo(view.label.replace('/', '-hitbox/'), 'init');
   await view.show();
-  await view.emit('complete-setup', height);
 }
 
 export function App() {
   const structure = useSelector(Selectors.placeholder);
-  const height = useSelector(Selectors.settings.height);
 
   useEffect(() => {
-    onMount(height);
+    onMount();
   }, []);
 
   if (!structure) {