+++ src/background/monitor.rs
@@ -1,8 +1,14 @@
 use color_eyre::eyre::eyre;
 use getset::{Getters, MutGetters};
-use windows::Win32::Graphics::Gdi::HMONITOR;
 
-use crate::{error_handler::Result, seelen_bar::FancyToolbar, seelen_weg::SeelenWeg, state::State};
+use crate::{
+    error_handler::Result,
+    seelen_bar::FancyToolbar,
+    seelen_weg::SeelenWeg,
+    state::State,
+};
+
+use windows::Win32::Graphics::Gdi::HMONITOR;
 
 #[derive(Getters, MutGetters)]
 pub struct Monitor {
@@ -10,7 +16,7 @@ pub struct Monitor {
     hmonitor: HMONITOR,
     #[getset(get = "pub", get_mut = "pub")]
     toolbar: Option<FancyToolbar>,
-    #[allow(dead_code)]
+    #[getset(get = "pub", get_mut = "pub")]
     weg: Option<SeelenWeg>,
 }
 
@@ -30,6 +36,13 @@ impl Monitor {
             match FancyToolbar::new(hmonitor.0) {
                 Ok(bar) => monitor.toolbar = Some(bar),
                 Err(e) => log::error!("Failed to create Toolbar: {}", e),
+            }
+        }
+
+        if settings.is_weg_enabled() {
+            match SeelenWeg::new(hmonitor.0) {
+                Ok(weg) => monitor.weg = Some(weg),
+                Err(e) => log::error!("Failed to create SeelenWeg: {}", e),
             }
         }