+++ src/background/seelen_weg/handler.rs
@@ -15,11 +15,11 @@ use super::SeelenWeg;
 
 #[command]
 pub fn enum_opened_apps() {
-    std::thread::spawn(|| {
+    /* std::thread::spawn(|| {
         if let Some(weg) = SEELEN.lock().weg() {
             weg.update_ui();
         }
-    });
+    }); */
 }
 
 #[derive(Deserialize)]